Change package from com.thefuntasty.hauler -> app.futured.hauler
