+++ core/src/main/AndroidManifest.xml
@@ -1,1 +1,1 @@
-<manifest package="com.thefuntasty.hauler"/>
+<manifest package="app.futured.hauler"/>
