+++ core/src/main/kotlin/app/futured/hauler/DragDirection.kt
@@ -1,0 +1,5 @@
+package app.futured.hauler
+
+enum class DragDirection {
+    UP, DOWN
+}
