+++ core/src/main/kotlin/app/futured/hauler/OnDragActivityListener.kt
@@ -1,0 +1,5 @@
+package app.futured.hauler
+
+interface OnDragActivityListener {
+    fun onDrag(elasticOffsetPixels: Float, rawOffset: Float)
+}
