+++ core/src/main/kotlin/app/futured/hauler/OnDragDismissedListener.kt
@@ -1,0 +1,5 @@
+package app.futured.hauler
+
+interface OnDragDismissedListener {
+    fun onDismissed(dragDirection: DragDirection)
+}
