--- core/src/main/kotlin/com/thefuntasty/hauler/ColorUtils.kt
@@ -1,13 +1,0 @@
-package com.thefuntasty.hauler
-
-import androidx.annotation.CheckResult
-import androidx.annotation.ColorInt
-import androidx.annotation.IntRange
-
-internal object ColorUtils {
-
-    @CheckResult
-    @ColorInt
-    fun modifyAlpha(@ColorInt color: Int, @IntRange(from = 0, to = 255) alpha: Int): Int =
-        color and 0x00ffffff or (alpha shl 24)
-}
