--- core/src/main/kotlin/com/thefuntasty/hauler/DragDirection.kt
@@ -1,5 +1,0 @@
-package com.thefuntasty.hauler
-
-enum class DragDirection {
-    UP, DOWN
-}
