--- core/src/main/kotlin/com/thefuntasty/hauler/OnDragActivityListener.kt
@@ -1,5 +1,0 @@
-package com.thefuntasty.hauler
-
-interface OnDragActivityListener {
-    fun onDrag(elasticOffsetPixels: Float, rawOffset: Float)
-}
