--- core/src/main/kotlin/com/thefuntasty/hauler/OnDragDismissedListener.kt
@@ -1,5 +1,0 @@
-package com.thefuntasty.hauler
-
-interface OnDragDismissedListener {
-    fun onDismissed(dragDirection: DragDirection)
-}
