+++ databinding/src/main/AndroidManifest.xml
@@ -1,1 +1,1 @@
-<manifest package="com.thefuntasty.hauler.databinding"/>
+<manifest package="app.futured.hauler.databinding"/>
