+++ databinding/src/main/kotlin/app/futured/hauler/databinding/IsDragEnabledAdapter.kt
@@ -1,0 +1,9 @@
+package app.futured.hauler.databinding
+
+import androidx.databinding.BindingAdapter
+import app.futured.hauler.HaulerView
+
+@BindingAdapter("app:isDragEnabled")
+fun HaulerView.isDragEnabled(isDragEnabled: Boolean) {
+    this.setDragEnabled(isDragEnabled)
+}
