+++ databinding/src/main/kotlin/app/futured/hauler/databinding/IsDragUpEnabledAdapter.kt
@@ -1,0 +1,9 @@
+package app.futured.hauler.databinding
+
+import androidx.databinding.BindingAdapter
+import app.futured.hauler.HaulerView
+
+@BindingAdapter("app:isDragUpEnabled")
+fun HaulerView.isDragUpEnabled(isDragUpEnabled: Boolean) {
+    this.setDragUpEnabled(isDragUpEnabled)
+}
