+++ databinding/src/main/kotlin/app/futured/hauler/databinding/IsScrollableAdapter.kt
@@ -1,0 +1,9 @@
+package app.futured.hauler.databinding
+
+import androidx.databinding.BindingAdapter
+import app.futured.hauler.LockableNestedScrollView
+
+@BindingAdapter("app:isScrollable")
+fun LockableNestedScrollView.isScrollable(isScrollable: Boolean) {
+    setScrollEnabled(isScrollable)
+}
