--- databinding/src/main/kotlin/com/thefuntasty/hauler/databinding/IsDragEnabledAdapter.kt
@@ -1,9 +1,0 @@
-package com.thefuntasty.hauler.databinding
-
-import androidx.databinding.BindingAdapter
-import com.thefuntasty.hauler.HaulerView
-
-@BindingAdapter("app:isDragEnabled")
-fun HaulerView.isDragEnabled(isDragEnabled: Boolean) {
-    this.setDragEnabled(isDragEnabled)
-}
