--- databinding/src/main/kotlin/com/thefuntasty/hauler/databinding/IsScrollableAdapter.kt
@@ -1,9 +1,0 @@
-package com.thefuntasty.hauler.databinding
-
-import androidx.databinding.BindingAdapter
-import com.thefuntasty.hauler.LockableNestedScrollView
-
-@BindingAdapter("app:isScrollable")
-fun LockableNestedScrollView.isScrollable(isScrollable: Boolean) {
-    setScrollEnabled(isScrollable)
-}
