+++ sample/src/main/kotlin/app/futured/haulersample/draggable/databinding/DatabindingActivityState.kt
@@ -1,0 +1,5 @@
+package app.futured.haulersample.draggable.databinding
+
+data class DatabindingActivityState(
+    val isDragEnabled: Boolean
+)
