+++ sample/src/main/kotlin/app/futured/haulersample/draggable/databinding/DatabindingActivityView.kt
@@ -1,0 +1,7 @@
+package app.futured.haulersample.draggable.databinding
+
+import app.futured.hauler.DragDirection
+
+interface DatabindingActivityView {
+    fun onDragDismissed(dragDirection: DragDirection)
+}
