--- sample/src/main/kotlin/com/thefuntasty/haulersample/draggable/databinding/DatabindingActivity.kt
@@ -1,41 +1,0 @@
-package com.thefuntasty.haulersample.draggable.databinding
-
-import android.content.Context
-import android.content.Intent
-import android.os.Bundle
-import androidx.appcompat.app.AppCompatActivity
-import androidx.databinding.DataBindingUtil
-import com.thefuntasty.hauler.DragDirection
-import com.thefuntasty.haulersample.R
-import com.thefuntasty.haulersample.databinding.ActivityDatabindingBinding
-import kotlinx.android.synthetic.main.activity_databinding.*
-
-class DatabindingActivity : AppCompatActivity(), DatabindingActivityView {
-    companion object {
-        fun getStartIntent(context: Context): Intent = Intent(context, DatabindingActivity::class.java)
-    }
-    lateinit var binding: ActivityDatabindingBinding
-
-    private var state = DatabindingActivityState(true)
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-
-        binding = DataBindingUtil.setContentView(this, R.layout.activity_databinding)
-        refreshBinding()
-
-        dragEnabledCheck.setOnCheckedChangeListener { _, isChecked ->
-            state = state.copy(isDragEnabled = isChecked)
-            refreshBinding()
-        }
-    }
-
-    private fun refreshBinding() {
-        binding.viewState = state
-        binding.view = this
-        binding.executePendingBindings()
-    }
-
-    override fun onDragDismissed(dragDirection: DragDirection) {
-        finish()
-    }
-}
