--- sample/src/main/kotlin/com/thefuntasty/haulersample/draggable/databinding/DatabindingActivityState.kt
@@ -1,5 +1,0 @@
-package com.thefuntasty.haulersample.draggable.databinding
-
-data class DatabindingActivityState(
-    val isDragEnabled: Boolean
-)
