--- sample/src/main/kotlin/com/thefuntasty/haulersample/draggable/databinding/DatabindingActivityView.kt
@@ -1,7 +1,0 @@
-package com.thefuntasty.haulersample.draggable.databinding
-
-import com.thefuntasty.hauler.DragDirection
-
-interface DatabindingActivityView {
-    fun onDragDismissed(dragDirection: DragDirection)
-}
