+++ sample/src/main/res/layout/activity_advanced.xml
@@ -1,9 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
-<com.thefuntasty.hauler.HaulerView
+<app.futured.hauler.HaulerView
         xmlns:android="http://schemas.android.com/apk/res/android"
         xmlns:tools="http://schemas.android.com/tools"
         xmlns:app="http://schemas.android.com/apk/res-auto"
-        tools:context=".draggable.advanced.AdvancedActivity"
+        tools:context="app.futured.haulersample.draggable.advanced.AdvancedActivity"
         android:id="@+id/advancedHaulerView"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
@@ -14,7 +14,7 @@
         app:fadeSystemBars="false"
         app:dragDismissDistance="200dp">
 
-    <com.thefuntasty.hauler.LockableNestedScrollView
+    <app.futured.hauler.LockableNestedScrollView
             android:id="@+id/scrollViewParent"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
@@ -32,7 +32,7 @@
                     android:layout_gravity="center"
                     android:text="@string/swipe_to_dismiss"/>
 
-            <com.thefuntasty.haulersample.draggable.advanced.IgnoredAreaView
+            <app.futured.haulersample.draggable.advanced.IgnoredAreaView
                     android:id="@+id/ignoredAreaView"
                     android:layout_width="200dp"
                     android:layout_marginTop="20dp"
@@ -44,7 +44,7 @@
 
         </LinearLayout>
 
-    </com.thefuntasty.hauler.LockableNestedScrollView>
+    </app.futured.hauler.LockableNestedScrollView>
 
-</com.thefuntasty.hauler.HaulerView>
+</app.futured.hauler.HaulerView>
